+++ advanced_samples/Grafana/src/types.ts
@@ -1,20 +1,25 @@
 import { DataQuery, DataSourceJsonData } from '@grafana/data';
 
 export enum SdsDataSourceType {
-  OCS = 'ocs',
-  EDS = 'eds',
+  OCS = 'OCS',
+  EDS = 'EDS',
 }
 
 export interface SdsQuery extends DataQuery {
-  namespace: string;
   stream: string;
 }
 
 export interface SdsDataSourceOptions extends DataSourceJsonData {
   type: SdsDataSourceType;
-  port: string;
-  url: string;
-  version: string;
-  tenant: string;
-  client: string;
+  eds_port: string;
+  ocs_url: string;
+  ocs_version: string;
+  ocs_tenant: string;
+  ocs_client: string;
+  oauthPassThru: boolean;
+  namespace: string;
+}
+
+export interface SdsDataSourceSecureOptions {
+  ocs_secret: string;
 }